Refactor components to server components as possible
