+++ components/buttons/GoogleAuthButton.tsx
@@ -1,0 +1,29 @@
+'use client';
+
+import { auth } from '@/lib/firebase';
+import { GoogleAuthProvider, signInWithPopup } from 'firebase/auth';
+import { useRouter } from 'next/navigation';
+
+export const GoogleAuthButton = ({ children }: { children: React.ReactNode }) => {
+  const router = useRouter();
+
+  const handleGoogleSignIn = async () => {
+    const provider = new GoogleAuthProvider();
+    try {
+      await signInWithPopup(auth, provider);
+      router.push('/');
+    } catch (error) {
+      console.error('Error during Google sign-in: ', error);
+    }
+  };
+
+  return (
+    <button
+      type="button"
+      className="flex items-center justify-center size-[37.5px] transition-all duration-200 ease-linear p-0 text-white btn bg-orange-500 border-orange-500 hover:text-white hover:bg-orange-600 hover:border-orange-600 focus:text-white focus:bg-orange-600 focus:border-orange-600 active:text-white active:bg-orange-600 active:border-orange-600"
+      onClick={handleGoogleSignIn}
+    >
+      {children}
+    </button>
+  );
+};
